mlfqs

+++ include/threads/thread.h
@@ -90,8 +90,8 @@ struct thread {
 	int priority;                       /* Priority. */
     int original_priority;
     
-    int nice;
-    int recent_cpu;
+    int32_t nice;
+    int64_t recent_cpu;
 	/* Shared between thread.c and synch.c. */
 	struct list_elem elem;/* List element. */
     struct lock * want_lock;
